feat: Remove extra 'user_list' call in ogdch_user_list
Previously, we got all users from the database with a query, and then called the action 'user_list' to get a filtered, ordered list of usernames. This made a separate database query. We then used the list of usernames to filter and order the list of users.

By combining the filtering and ordering in the 'user_list' action with our simpler database query, we remove the need to call 'user_list' separately.